+++ packages/angular/avatar/public-api.ts
@@ -1,1 +1,7 @@
 export * from './avatar';
+export {
+  HELL_AVATAR_GROUP_DIRECTIVES,
+  HellAvatarGroup,
+  HellAvatarGroupItem,
+  HellAvatarGroupOverflow,
+} from './avatar-group';